Intermediate commit: filtering files by mask/extension was replaced with regexp

+++ src/main/java/org/psp/recognition/fs/FSDirectory.java
@@ -4,25 +4,28 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FilenameFilter;
+import java.util.ArrayList;
 
 public class FSDirectory extends AbstractFSObject {
     private final static Logger LOG = LoggerFactory.getLogger(FSDirectory.class.getName());
     private boolean isToDelete = false;
-    private String nameMask = null;
-    private String extensionMask = null;
-    private int level = 1; // with subdirectories
+    private ArrayList<String> namePatterns;
+    private boolean withSubdirectories = false; // with subdirectories
 
     public FSDirectory(String pathname, boolean isToDelete) {
         super(pathname);
         this.isToDelete = isToDelete;
     }
 
-    public FSDirectory(String pathname, String nameMask, String extensionMask, int level) {
+    public FSDirectory(boolean withSubdirectories, String pathname) {
         super(pathname);
-        this.nameMask = nameMask;
-        this.extensionMask = extensionMask;
-        this.level = level;
+        this.withSubdirectories = withSubdirectories;
+    }
+
+    public FSDirectory(String pathname, ArrayList<String> namePatterns, boolean withSubdirectories) {
+        super(pathname);
+        this.namePatterns = namePatterns;
+        this.withSubdirectories = withSubdirectories;
     }
 
     public FSDirectory(String pathname) {
@@ -30,26 +33,27 @@ public class FSDirectory extends AbstractFSObject {
     }
 
     public boolean iterate() {
-        FSFileNameFilter fsFileNameFilter = new FSFileNameFilter(nameMask, extensionMask);
+        FSFileNameFilter fsFileNameFilter = new FSFileNameFilter(namePatterns);
         File[] fileList = this.listFiles(fsFileNameFilter);
-        LOG.debug("fileList = ", fileList);
-        for (File fsItem : fileList) {
-            LOG.debug("fsItem = {}", fsItem);
+        if (fileList != null) {
+            for (File fsItem : fileList) {
+                LOG.debug("fsItem = {}", fsItem);
 
-            if (fsItem.isDirectory() && level == 1) {
-                LOG.debug("Listing directory {}", fsItem);
-                FSDirectory directory = new FSDirectory(fsItem.getPath(), nameMask, extensionMask, level);
-                if (!directory.iterate())
-                    throw new IllegalStateException("Can not list " + directory);
-            } else {
-                LOG.debug("directory = {}", this.getPath());
-                FSFile file = new FSFile(fsItem.getPath());
-                if (!file.run())
-                    throw new IllegalStateException("Can not run " + file);
-            }
-            if (isToDelete) {
-                LOG.debug("Delete fsItem {}", fsItem);
-                this.delete();
+                if (fsItem.isDirectory() && withSubdirectories) {
+                    LOG.debug("Listing directory {}", fsItem);
+                    FSDirectory directory = new FSDirectory(fsItem.getPath(), namePatterns, withSubdirectories);
+                    if (!directory.iterate())
+                        throw new IllegalStateException("Can not list " + directory);
+                } else {
+                    LOG.debug("directory = {}", this.getPath());
+                    FSFile file = new FSFile(fsItem.getPath());
+                    if (!file.run())
+                        throw new IllegalStateException("Can not run " + file);
+                }
+                if (isToDelete) {
+                    LOG.debug("Delete fsItem {}", fsItem);
+                    this.delete();
+                }
             }
         }
         return true;
